Added basic install & migration generators.
I did this because it is a total pain in the ass to have to create your own
migration files. This helps the users make sure that their migrations class
names and file names are in sync, that they are ordered correctly by using
timestamp of creation, as well as making sure the migraiton inherits properly
from the Cequel::Migration base class that will soon exist and provide the
basic functionality for making the connection to the cequel and allowing you
to make changes.