Switched the schema_migrations version to bigint.
I did this because I wanted to be able to take timestamps in the form of
YYYYMMDDHHMMSS as a string, be able to convert that to an int and use that as
the migration version. The normal int type didn't allow for values that large.
Hence, the switch to the bigint data type.