Get rid of sudo cqlsh & add keyspace.
There is no reason that 'sudo' should be require to make a cqlsh connection to
a Cassandra node. If it is there is a problem somewhere higher up and
permissions at the system level are incorrect for the installation. Therefore,
I removed the forced 'sudo' out of the cequel:shell task. Also, the task was
NOT specify the keyspace that was defined in the cequel.yml. This is important
because if you have perms setup on a per keyspace basis you only want to have
to have the credentials necessary for that keyspace.